requirements.txt
asgiref==3.5.2
autopep8==1.7.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.1
django-crispy-forms==1.14.0
django-import-export==2.8.0
django-safedelete==1.3.0
django-simple-history==3.1.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
pycodestyle==2.9.1
pytz==2022.2.1
PyYAML==6.0
sqlparse==0.4.2
tablib==3.2.1
toml==0.10.2
xlrd==2.0.1
xlwt==1.3.0
